removed contacts page and added footer to remaining pages
