+++ src/components/index.ts
@@ -1,10 +1,10 @@
-export { Button } from './button';
-export { ContactInfo } from './contact-info';
-export { EventInfo } from './event-info';
-export { Header } from './header';
-export { Headline } from './headline';
-export { Logo } from './logo';
-export { NavBar } from './nav-bar';
-export { NavItem } from './nav-item';
-export { SponsorInfo } from './sponsor-info';
-export { default as ScrollToTop } from './scroll-to-top';
+export { Button } from "./button";
+export { EventInfo } from "./event-info";
+export { Header } from "./header";
+export { Headline } from "./headline";
+export { Logo } from "./logo";
+export { NavBar } from "./nav-bar";
+export { NavItem } from "./nav-item";
+export { SponsorInfo } from "./sponsor-info";
+export { default as ScrollToTop } from "./scroll-to-top";
+export { Footer } from "./footer";
