+++ src/layouts/index.ts
@@ -1,4 +1,3 @@
-export { ContactUs } from './contact-us';
-export { Events } from './events';
-export { Home } from './home';
-export { Sponsors } from './sponsors';
+export { Events } from "./events";
+export { Home } from "./home";
+export { Sponsors } from "./sponsors";
