+++ src/routes.ts
@@ -1,30 +1,19 @@
-import {
-  ContactUs,
-  Events,
-  Home,
-  Sponsors,
-} from "./layouts"
-
+import { Events, Home, Sponsors } from "./layouts";
 
 export const routes = [
   {
     component: Home,
-    path: '/',
-    text: 'Home',
+    path: "/",
+    text: "Home",
   },
   {
     component: Events,
-    path: '/events',
-    text: 'Events',
+    path: "/events",
+    text: "Events",
   },
   {
     component: Sponsors,
-    path: '/sponsors',
-    text: 'Sponsors',
-  },
-  {
-    component: ContactUs,
-    path: '/contact-us',
-    text: 'Contact Us',
+    path: "/sponsors",
+    text: "Sponsors",
   },
 ];
